requirements.txt
python-telegram-bot==20.8
apscheduler>=3.10.4
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
certifi>=2023.7.22
